+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2018-07-30
+## [0.1.1] - 2018-07-30
+- Correct security with cgroup namespaces.
+- Interactive mode (for easily landing into a shell).
+- Environment variables forwarding and passing.
+
+## [0.1.0] - 2018-07-30
 ### Added
 - A basic sandbox using linux namespaces and the cgroups subsystem for limits
   and isolation.